Added some textures and applied wood texture to the cube
The textures are from https://ambientcg.com/ where it is stated that:

"All the PBR materials, brushes, photos and 3D models which are offered for download on ambientCG are provided under the Creative Commons CC0 1.0 Universal License."

The CC0 1.0 states:

"The person who associated a work with this deed has dedicated the work to the public domain by waiving all of his or her rights to the work worldwide under copyright law, including all related and neighboring rights, to the extent allowed by law.

You can copy, modify, distribute and perform the work, even for commercial purposes, all without asking permission. See Other Information below."

So... no problems with copyright or whatever :)