FINALLY FIXED DIFFUSE LIGHTING
The problem was that I set the vertex buffer attribute indexes in the Renderer header file in the class declaration but changed them in the constructor definition. Because of this there were no normals being passed properly to the vertex shader, the shader was getting texture coordinates instead of normals.