It works, I don't know why, I also don't know why it didn't work
The light source object was not being rendered. I made another branch to compare with previous versions, didn't find the problem. Reverted to current branch problem was still there. I commented out a line in a function that's not being called (checked this with debugger) and it suddenly was being rendered again..............